There are two kinds of input neurons : direction and amplitude.

+++ src/brain.hpp
@@ -10,8 +10,8 @@ class Brain
 {
     private:
         neural::Network m_net;
-        std::vector<size_t> m_input_channels;
-        std::vector<size_t> m_output_channels;
+        std::vector<int> m_input_channels;
+        std::vector<int> m_output_channels;
         size_t m_hard_input_count;
         size_t m_hard_output_count;
 
@@ -21,8 +21,8 @@ class Brain
               const std::vector<std::string> &hard_outputs);
         std::vector<double> answer(std::vector<double> inputs);
         void punish(std::vector<double> errors);
-        std::vector<size_t> get_input_channels();
-        std::vector<size_t> get_output_channels();
+        std::vector<int> get_input_channels();
+        std::vector<int> get_output_channels();
 };
 
 #endif // ALIFE_BRAIN_HPP
